perf: resize static image assets and use loading='lazy'
these are for now just stub assets. we will also need to figure out how
to properly use the sanity image cdn to serve optimized image assets
responsively based on screen size. it is already there, we just need to
use it. for now, quick fix for these assets which are very large and make
the page seem slower to load.